refactor: 옷장, 캘린더 로직 리팩토링
