--- app/src/main/java/com/example/onfit/BestItem.kt
@@ -1,7 +1,0 @@
-package com.example.onfit.data.model
-
-data class BestItem(
-    val imageResId: Int,   // drawable 리소스 ID
-    val rank: String,
-    val name: String
-)
