--- app/src/main/java/com/example/onfit/CommunityItem.kt
@@ -1,7 +1,0 @@
-package com.example.onfit
-
-data class CommunityItem(
-    val imageResId: Int,
-    val nickname: String,
-    val likeCount: Int
-)
