--- app/src/main/java/com/example/onfit/MainActivity.kt
@@ -1,30 +1,0 @@
-package com.example.onfit
-
-import android.os.Bundle
-import androidx.activity.enableEdgeToEdge
-import androidx.appcompat.app.AppCompatActivity
-import androidx.core.view.ViewCompat
-import androidx.core.view.WindowInsetsCompat
-import androidx.navigation.fragment.NavHostFragment         // ✅ 여기에 주목
-import androidx.navigation.ui.setupWithNavController
-import com.example.onfit.databinding.ActivityMainBinding
-
-class MainActivity : AppCompatActivity() {
-
-    private lateinit var binding: ActivityMainBinding
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        enableEdgeToEdge()
-        binding = ActivityMainBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-
-        // NavHostFragment를 직접 가져와서 NavController 얻기
-        val navHostFragment = supportFragmentManager
-            .findFragmentById(R.id.nav_host_fragment) as NavHostFragment
-        val navController = navHostFragment.navController
-
-        // BottomNavigationView와 NavController를 연결 → 자동으로 Fragment 전환
-        binding.bottomNavigationView.setupWithNavController(navController)
-    }
-}
