--- app/src/main/java/com/example/onfit/OutfitItem.kt
@@ -1,6 +1,0 @@
-package com.example.onfit
-
-data class OutfitItem(
-    val imageResId: Int,
-    var isClosetButtonActive: Boolean = true // 아이템이 옷장에 있는지
-)
