--- app/src/main/java/com/example/onfit/SimItem.kt
@@ -1,6 +1,0 @@
-package com.example.onfit.data.model
-
-data class SimItem(
-    val imageResId: Int,   // drawable 리소스 ID
-    val date: String       // 날씨 어떤지(많이 추움), 날짜 표시
-)
