+++ app/src/main/java/com/example/onfit/Wardrobe/fragment/ClothesDetailFragment.kt
@@ -1,4 +1,4 @@
-package com.example.onfit
+package com.example.onfit.Wardrobe.fragment
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -8,6 +8,7 @@ import android.widget.*
 import android.widget.LinearLayout
 import androidx.fragment.app.Fragment
 import androidx.navigation.fragment.findNavController
+import com.example.onfit.R
 
 class ClothesDetailFragment : Fragment() {
