+++ app/src/main/java/com/example/onfit/calendar/CalendarFragment.kt
@@ -1,4 +1,4 @@
-package com.example.onfit
+package com.example.onfit.calendar
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,11 +7,11 @@ import android.view.ViewGroup
 import android.widget.TextView
 import android.widget.Toast
 import androidx.fragment.app.Fragment
-import androidx.navigation.NavOptions
 import androidx.navigation.fragment.findNavController
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.PagerSnapHelper
 import androidx.recyclerview.widget.RecyclerView
+import com.example.onfit.R
 import java.util.*
 
 class CalendarFragment : Fragment() {
