+++ app/src/main/res/layout/fragment_wardrobe.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".WardrobeFragment"
+    tools:context=".Wardrobe.fragment.WardrobeFragment"
     android:backgroundTint="@color/white">
 
     <!-- 상단 카테고리 -->
