+++ app/src/main/res/navigation/nav_graph.xml
@@ -12,12 +12,12 @@
 
     <fragment
         android:id="@+id/wardrobeFragemnt"
-        android:name="com.example.onfit.WardrobeFragment"
+        android:name="com.example.onfit.Wardrobe.fragment.WardrobeFragment"
         android:label="옷장"/>
 
     <fragment
         android:id="@+id/calendarFragemnt"
-        android:name="com.example.onfit.CalendarFragment"
+        android:name="com.example.onfit.calendar.CalendarFragment"
         android:label="캘린더"/>
 
     <fragment
@@ -32,17 +32,17 @@
 
     <fragment
         android:id="@+id/styleOutfitsFragment"
-        android:name="com.example.onfit.StyleOutfitsFragment"
+        android:name="com.example.onfit.Wardrobe.fragment.StyleOutfitsFragment"
         android:label="스타일별 Outfit"/>
 
     <fragment
         android:id="@+id/wardrobeSearchFragment"
-        android:name="com.example.onfit.WardrobeSearchFragment"
+        android:name="com.example.onfit.Wardrobe.fragment.WardrobeSearchFragment"
         android:label="WardrobeSearchFragment" />
 
     <fragment
         android:id="@+id/clothesDetailFragment"
-        android:name="com.example.onfit.ClothesDetailFragment"
+        android:name="com.example.onfit.Wardrobe.fragment.ClothesDetailFragment"
         android:label="ClothesDetailFragment"
         tools:layout="@layout/fragment_clothes_detail">
         <argument
@@ -55,7 +55,7 @@
 
     <fragment
         android:id="@+id/addItemFragment"
-        android:name="com.example.onfit.AddItemFragment"
+        android:name="com.example.onfit.Wardrobe.fragment.AddItemFragment"
         android:label="AddItemFragment"
         tools:layout="@layout/fragment_add_item">
         <argument
@@ -71,7 +71,7 @@
 
     <fragment
         android:id="@+id/addItemFragment"
-        android:name="com.example.onfit.AddItemFragment"
+        android:name="com.example.onfit.Wardrobe.fragment.AddItemFragment"
         android:label="AddItemFragment"
         tools:layout="@layout/fragment_add_item">
         <argument
